requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
wandb>=0.15.0
pydicom>=2.3.0  # Optional, for DICOM support
tqdm>=4.65.0
scikit-learn>=1.2.0
pillow>=9.5.0
